\name{gendoo.Hs.db}
\alias{gendoo.Hs.db}
\alias{gendoo.Hs}
\title{Gendoo Human Database package}
\description{
  
  Welcome to the gendoo.Hs.db annotation package.  The purpose of
  this package is to provide mapping table for Entrez Gene ID and
  MeSH ID. The original data are provided at http://gendoo.dbcls.jp/.
  
  You can learn what objects this package supports with the following command:

  \code{ls("package:gendoo.Hs.db")} 
  
  Each of these objects has their own manual page detailing where
  relevant data was obtained along with some examples of how to use it.
}
\examples{
  ls("package:gendoo.Hs.db")
}
\keyword{datasets}
